Setup
* .travis.yml: Changed python versions
* tests.test_timeseries: Moved from ./
* tox.ini: Changed versions
