+++ .travis.yml
@@ -4,7 +4,7 @@ language: python
 python:
   - 3.8
   - 3.7
-  - 3.6
+  - 3.9
 
 # Command to install dependencies, e.g. pip install -r requirements.txt --use-mirrors
 install: pip install -U tox-travis
